+++ packages/system/src/collections/index.ts
@@ -1,10 +1,8 @@
-export * from './apiKey'
 export * from './file'
 export * from './log'
 export * from './resourceUsage'
 export { user } from './user'
 
-export type { ApiKey } from './apiKey/description'
 export type { File } from './file/description'
 export type { Log } from './log/description'
 export type { ResourceUsage } from './resourceUsage/description'