+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sonata-api/types
 
+## 0.1.19
+
+### Patch Changes
+
+- Improve DecodedToken
+
 ## 0.1.18
 
 ### Patch Changes